Implement Markdown trailing space line break preservation
Markdown uses two or more trailing spaces on a line to indicate a forced
line break `<br/>` - these will be preserved for files with a markdown
extension (default = `.md` or `.markdown`).
Add `--markdown-linebreak-ext=X,Y` to set extensions (`*` matches any),
and `--no-markdown-linebreak-ext` to disable this feature.

Update README.md to include information on these arguments as well as
arguments added to other hooks